Address PR #229 review: cache stamp, poller test, DeviceState PercentCompleted
- **Camera DeviceState** (AlpacaCore, all vendors): emit PercentCompleted
  as an integer (ICameraV4 short); as a double ConformU could not coerce
  it and logged "not included in the DeviceState response" on every
  camera log.
- **ToupTek Camera Driver** (AlpacaCore): the getters' live-read fallback
  now stamps the thermal cache so a stalled poller doesn't turn every
  call into an SDK read (review note).
- **ToupTek Unit Tests**: fake SDK gains a holdable wait_image and a
  thread-safe call counter; new case asserts the poller primes the cache
  at connect, issues no thermal SDK call while an exposure is in flight
  (getters still answer from the cache), and resumes afterwards.
- **Docs**: CHANGELOG entries, AGENTS.md DeviceState value-type rule.

+++ AlpacaCore/include/alpacacore/camera_driver.h
@@ -89,7 +89,10 @@ public:
         add("HeatSinkTemperature", [this] { return get_heat_sink_temperature(); });
         add("ImageReady", [this] { return get_image_ready(); });
         add("IsPulseGuiding", [this] { return get_is_pulse_guiding(); });
-        add("PercentCompleted", [this] { return get_percent_completed(); });
+        // ICameraV4.PercentCompleted is a short: emit an integer, matching the
+        // GET endpoint. As a double ("0.0") ConformU cannot coerce it and
+        // reports the property "not included in the DeviceState response".
+        add("PercentCompleted", [this] { return static_cast<std::int32_t>(get_percent_completed()); });
 
         state.push_back({"TimeStamp", device_state_timestamp()});
         return state;
